Change postive int arguments to use size_t in prototypes.
cppad_swig.omh: add \B latex macro.
a_fun.cpp: Add documentation for forward.
forward_xam.xam: extend to include second order example.
cplusplus.m4: add a_fun to using commands at beginning of function.

+++ include/cppad/swig/a_fun.hpp
@@ -47,7 +47,7 @@ class CPPAD_SWIG_LIB_PUBLIC a_fun
 	// constrtuctor
 	a_fun( const std::vector<a_double>& ax, const std::vector<a_double>& ay );
 	// forward
-	std::vector<double> forward(int p, const std::vector<double>& xp );
+	std::vector<double> forward(size_t p, const std::vector<double>& xp );
 };
 
 } // END_CPPAD_SWIG_NAMESPACE
